backend/requirements.txt
asgiref==3.11.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
django==5.2.8
django-allauth==65.13.0
django-filter==25.2
django-redis==6.0.0
django-silk==5.4.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
docker==7.1.0
google-auth==2.43.0
google-auth-oauthlib==1.2.2
gprof2dot==2025.4.14
idna==3.11
iniconfig==2.3.0
oauthlib==3.3.1
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1-modules==0.4.2
pygments==2.19.2
pyjwt==2.10.1
pytest==9.0.1
python-dotenv==1.2.1
redis==7.0.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
ruff==0.14.5
sqlparse==0.5.3
testcontainers==4.13.3
typing-extensions==4.15.0
urllib3==2.5.0
wrapt==2.0.1
